Add noop escape to weapons

+++ src/database/character_diff/weapons.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use eyre::Result;
 use sqlx::{query, Postgres, Transaction};
 
@@ -8,24 +10,53 @@ use crate::{
 
 #[derive(Debug, Default)]
 pub struct WeaponsDiff {
+    noop: bool,
     created_weapons: Vec<(Weapon, Option<EquipHand>)>,
     owned_weapons: Vec<(i32, Option<EquipHand>)>,
 }
 
 impl Weapons {
     pub fn compare_newer(&self, newer: &Self) -> WeaponsDiff {
-        let mut diff = WeaponsDiff::default();
+
+        let mut noop = true;
+        let mut created_weapons = Vec::new();
+        let mut new_owned_weapons = HashMap::new();
 
         for (_, maybe_equip_hand, weapon) in newer.iter() {
             if let Some(id) = weapon.id() {
-                diff.owned_weapons.push((id, maybe_equip_hand));
+                new_owned_weapons.insert(id, maybe_equip_hand);
             } else {
-                diff.created_weapons
-                    .push((weapon.clone(), maybe_equip_hand));
+                noop = false;
+                created_weapons.push((weapon.clone(), maybe_equip_hand));
             }
         }
 
-        diff
+        if noop {
+            let mut old_owned_weapons = HashMap::new();
+
+            for (_, maybe_equip_hand, weapon) in self.iter() {
+                if let Some(id) = weapon.id() {
+                    old_owned_weapons.insert(id, maybe_equip_hand);
+                }
+            }
+
+            noop = old_owned_weapons != new_owned_weapons;
+        }
+
+
+        if noop {
+            WeaponsDiff {
+                noop,
+                created_weapons: Vec::new(),
+                owned_weapons: Vec::new(),
+            }
+        } else {
+            WeaponsDiff {
+                noop,
+                created_weapons,
+                owned_weapons: new_owned_weapons.into_iter().collect()
+            }
+        }
     }
 }
 
@@ -35,6 +66,10 @@ impl WeaponsDiff {
         transaction: &mut Transaction<'_, Postgres>,
         character_id: i32,
     ) -> Result<()> {
+        if self.noop {
+            return Ok(());
+        }
+
         // Drop all owned/equipped records
         query!(
             "DELETE FROM character_weapons
